requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
Pillow==10.3.0
easyocr==1.7.1
pytesseract==0.3.10
pyzbar==0.1.9
